Combi phototransistor sensor array
Combined sensor array based on Osram SFH3710-4 phototransistor along with Schmitt inverters and shunts on the same board. This connects directly to the target interface. Optical characteristics will be adequately consistent using the "-4" suffix (high sensitivity bin) phototransistors. The numbers looked dodgy with the full spread of sensitivity values though.  The -Z or no suffix option is cheaper, but there's a greater chance of the system not working due to sensitivity spread and you'll need to get more photrans to swap for any outliers.